+++ test/extras/test_logging.rb
@@ -7,17 +7,17 @@ require "logger"
 require "stringio"
 
 class TestLogging < Minitest::Test
-  def test_log_to_returns_self
+  def test_returns_logging_instance
     server = build_server
     logger = Logger.new(StringIO.new)
 
-    assert_equal server, server.log_to(logger)
+    assert_instance_of Reclamo::Extras::Logging, Reclamo::Extras::Logging.new(server, logger)
   end
 
   def test_logs_successful_response
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3],"id":1}')
 
     assert_match(/add/, output.string)
@@ -27,7 +27,7 @@ class TestLogging < Minitest::Test
   def test_logs_at_info_level_by_default
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3],"id":1}')
 
     assert_match(/INFO/, output.string)
@@ -36,7 +36,7 @@ class TestLogging < Minitest::Test
   def test_logs_at_custom_level
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output), level: :debug)
+    Reclamo::Extras::Logging.new(server, Logger.new(output), level: :debug)
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3],"id":1}')
 
     assert_match(/DEBUG/, output.string)
@@ -45,7 +45,7 @@ class TestLogging < Minitest::Test
   def test_logs_errors_at_error_level
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"nonexistent","id":1}')
 
     assert_match(/ERROR/, output.string)
@@ -56,7 +56,7 @@ class TestLogging < Minitest::Test
   def test_logs_duration
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3],"id":1}')
 
     assert_match(/\d+\.\d+ms/, output.string)
@@ -65,7 +65,7 @@ class TestLogging < Minitest::Test
   def test_logs_method_name
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"divide","params":[6,2],"id":1}')
 
     assert_match(/divide/, output.string)
@@ -74,7 +74,7 @@ class TestLogging < Minitest::Test
   def test_logs_progname
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3],"id":1}')
 
     assert_match(/Reclamo/, output.string)
@@ -83,7 +83,7 @@ class TestLogging < Minitest::Test
   def test_notifications_are_logged
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[2,3]}')
 
     assert_match(/add/, output.string)
@@ -92,7 +92,7 @@ class TestLogging < Minitest::Test
   def test_multiple_requests_each_logged
     output = StringIO.new
     server = build_server
-    server.log_to(Logger.new(output))
+    Reclamo::Extras::Logging.new(server, Logger.new(output))
     server.handle('{"jsonrpc":"2.0","method":"add","params":[1,2],"id":1}')
     server.handle('{"jsonrpc":"2.0","method":"divide","params":[6,3],"id":2}')
 
@@ -100,11 +100,11 @@ class TestLogging < Minitest::Test
     assert_match(/divide/, output.string)
   end
 
-  def test_chainable_with_other_setup
+  def test_works_alongside_server_setup
     server = Reclamo::Server.new
                             .expose(Calculator)
-                            .log_to(Logger.new(StringIO.new))
                             .expose_method("ping") { "pong" }
+    Reclamo::Extras::Logging.new(server, Logger.new(StringIO.new))
 
     assert_equal 3, server.size
   end